testing/mochitest/tests/MochiKit-1.4.2/tests/test_MochiKit-Signal.html

<html>
<head>
    <script type="text/javascript" src="../MochiKit/Base.js"></script>
    <script type="text/javascript" src="../MochiKit/Iter.js"></script>
    <script type="text/javascript" src="../MochiKit/DOM.js"></script>
    <script type="text/javascript" src="../MochiKit/Style.js"></script>
    <script type="text/javascript" src="../MochiKit/Signal.js"></script>    
    <script type="text/javascript" src="../MochiKit/Logging.js"></script>    
    <script type="text/javascript" src="SimpleTest/SimpleTest.js"></script>        
    <link rel="stylesheet" type="text/css" href="SimpleTest/test.css">

</head>
<body>

Please ignore this button: <input type="submit" id="submit" /><br />

<pre id="test">
<script type="text/javascript" src="test_Signal.js"></script>
<script type="text/javascript">
try {
        
    tests.test_Signal({ok:ok, is:is});
    ok(true, "test suite finished!");
    
} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok(false, s);

}
</script>
</pre>
</body>
</html>
